+++ src/views/SignUpPage/components/3_선호지역선택/RegionItem.tsx
@@ -1,8 +1,8 @@
 import { useRecoilState } from 'recoil';
 import { styled } from 'styled-components';
 
-import { IcCheckboxBlue, IcCheckboxGrey } from '../../@common/assets/icons';
-import { PREFER_REGION_MAX_COUNT, SELECT_ALL } from '../constants/constants';
+import { IcCheckboxBlue, IcCheckboxGrey } from '../../../@common/assets/icons';
+import { PREFER_REGION_MAX_COUNT, SELECT_ALL } from '../../constants/constants';
 
 import { preferRegionState } from '@/recoil/atoms/signUpState';
 interface RegionItemProps {
