+++ src/views/SignUpPage/components/@common/IntroductionTextArea.tsx
